+++ heimdallm/bifrosts/sql/tests/sql/select/test_postgres.py
@@ -9,7 +9,7 @@ from .utils import PermissiveConstraints
 @dialects("postgres")
 def test_type_cast(dialect: str, Bifrost: Type[Bifrost]):
     """Prove prefix type casts are allowed"""
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
 
     query = """
 select col::int from t1