+++ heimdallm/bifrosts/sql/tests/sql/select/test_select.py
@@ -30,14 +30,14 @@ def test_aliased_select_column(dialect: str, Bifrost: Type[Bifrost]):
         def select_column_allowed(self, column: FqColumn) -> bool:
             return column.name in {"t2.col"}
 
-    bifrost = Bifrost.mocked(MyConstraints())
+    bifrost = Bifrost.validation_only(MyConstraints())
     bifrost.traverse(query)
 
 
 def test_unqualified_columns():
     query = "select col from t1"
 
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
     with pytest.raises(exc.UnqualifiedColumn) as e:
         bifrost.traverse(query, autofix=False)
     assert e.value.column == "col"
@@ -55,7 +55,7 @@ ORDER BY payment_date DESC
 LIMIT 5;
     """
 
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
     with pytest.raises(exc.UnqualifiedColumn) as e:
         bifrost.traverse(query, autofix=False)
     assert e.value.column == "customer_id"
@@ -79,7 +79,7 @@ def test_escapes(dialect: str, Bifrost: Type[Bifrost], query):
         def condition_column_allowed(self, column: FqColumn) -> bool:
             return column.name in {"t1.id", "t2.jid", "t1.jid"}
 
-    bifrost = Bifrost.mocked(MyConstraints())
+    bifrost = Bifrost.validation_only(MyConstraints())
     bifrost.traverse(query)
 
 
@@ -97,10 +97,10 @@ def test_disallowed_select_column(dialect: str, Bifrost: Type[Bifrost]):
         def select_column_allowed(self, fq_column: FqColumn) -> bool:
             return fq_column.name in {"t2.col"}
 
-    bifrost = Bifrost.mocked(AllowColumnConstraints())
+    bifrost = Bifrost.validation_only(AllowColumnConstraints())
     bifrost.traverse(query)
 
-    bifrost = Bifrost.mocked(DenyColumnConstraints())
+    bifrost = Bifrost.validation_only(DenyColumnConstraints())
     with pytest.raises(exc.IllegalSelectedColumn) as excinfo:
         bifrost.traverse(query)
 
@@ -124,7 +124,7 @@ def test_parameterized_constraints(dialect: str, Bifrost: Type[Bifrost]):
                 )
             ]
 
-    bifrost = Bifrost.mocked(RequiredConstraints())
+    bifrost = Bifrost.validation_only(RequiredConstraints())
     with pytest.raises(exc.MissingParameterizedConstraint) as excinfo:
         bifrost.traverse(query)
 
@@ -158,7 +158,7 @@ def test_parameterized_constraints(dialect: str, Bifrost: Type[Bifrost]):
 def test_broken_query(dialect: str, Bifrost: Type[Bifrost], query):
     """tests that we raise an exception when we cannot parse the query"""
 
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
     with pytest.raises(exc.InvalidQuery):
         bifrost.traverse(query)
 
@@ -166,7 +166,7 @@ def test_broken_query(dialect: str, Bifrost: Type[Bifrost], query):
 @dialects()
 def test_select_column_arith(dialect: str, Bifrost: Type[Bifrost]):
     query = "select t1.col + 1 as plus_one from t1"
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
     bifrost.traverse(query)
 
 
@@ -180,7 +180,7 @@ def test_select_column_arith(dialect: str, Bifrost: Type[Bifrost]):
 )
 def test_select_expr(dialect: str, Bifrost: Type[Bifrost], query):
     """select a non-column expression"""
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
     bifrost.traverse(query)
 
 
@@ -195,7 +195,7 @@ def test_conflicting_validation(dialect: str, Bifrost: Type[Bifrost]):
         def allowed_joins(self) -> Sequence[JoinCondition]:
             return []
 
-    bifrost = Bifrost.mocked(MyConstraints())
+    bifrost = Bifrost.validation_only(MyConstraints())
     bifrost.traverse(query)
 
 
@@ -207,7 +207,7 @@ def test_count_disallowed_column(dialect: str, Bifrost: Type[Bifrost]):
         def select_column_allowed(self, fq_column: FqColumn) -> bool:
             return fq_column.name != "film_actor.film_id"
 
-    bifrost = Bifrost.mocked(GeneralConstraints())
+    bifrost = Bifrost.validation_only(GeneralConstraints())
 
     # a query that uses the disallowed column in a count is allowed
     query = """