+++ heimdallm/bifrosts/sql/tests/sql/select/test_values.py
@@ -8,7 +8,7 @@ from .utils import PermissiveConstraints
 
 @dialects()
 def test_is_not_null(dialect: str, Bifrost: Type[Bifrost]):
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
 
     query = """
 SELECT film.title
@@ -24,7 +24,7 @@ LIMIT 20;
 
 @dialects()
 def test_is_null(dialect: str, Bifrost: Type[Bifrost]):
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
 
     query = """
 SELECT
@@ -43,7 +43,7 @@ LIMIT 20;
 @dialects()
 def test_agg_function_modifier_query(dialect: str, Bifrost: Type[Bifrost]):
     """distinct can be added in front of an aggregate function"""
-    bifrost = Bifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.validation_only(PermissiveConstraints())
 
     query = """
 SELECT COUNT(DISTINCT film.film_id) AS family_movies_rented